requirements.txt
py-trello
tweepy
requests